add more sig quit test

+++ src/tst/tst-live-quit-common.h
@@ -17,7 +17,7 @@ struct tst_live_quit_child
 };
 
 extern uint8_t tst_live_quit_allow_clone;
-extern uint8_t tst_live_quit_allow_group;
+extern int32_t tst_live_quit_allow_group;
 
 void tst_live_quit_clone (uint8_t *stack, int32_t *ptid, int32_t *ctid,
 			  void (*fn) (void *), void *data);
